LayoutTests/platform/gtk/accessibility/custom-elements/hidden-button-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("item-1") is null
PASS accessibilityController.accessibleElementById("item-2").role is "AXRole: AXButton"
PASS accessibilityController.accessibleElementById("item-2").stringAttributeValue("AXInvalid") is "grammar"
FAIL accessibilityController.accessibleElementById("item-2").stringAttributeValue("AXKeyShortcutsValue") should be Shift + E. Was .
PASS accessibilityController.accessibleElementById("item-3").role is "AXRole: AXButton"
PASS accessibilityController.accessibleElementById("item-3").stringAttributeValue("AXInvalid") is "spelling"
FAIL accessibilityController.accessibleElementById("item-3").stringAttributeValue("AXKeyShortcutsValue") should be Command + T. Was .
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
